src/Services/Customer/Application/Queries/GetCustomerByIdQueryHandler.cs
using Domain.Models;
using Domain.Services;
using MediatR;

namespace Application.Queries
{
    public class GetCustomerByIdQueryHandler : IRequestHandler<GetCustomerByIdQuery, Customer>
    {
        private readonly ICustomerService _customerService;

        public GetCustomerByIdQueryHandler(ICustomerService customerService)
        {
            _customerService = customerService;
        }

        public Task<Customer> Handle(GetCustomerByIdQuery request, CancellationToken cancellationToken)
        {
            var customer = _customerService.GetByIdAsync(request.Id);
            return customer;
        }
    }
}
